feat: Add Middleware (#6)
Add a Middleware that logs an HTTP Request.

- Define the RequestLog struct and define fields of the HTTP Request
that need to be logged (ex.Method,URI,Host,etc..). Also, define a
time field to store the time the HTTP Request was made.

- Implement createLogFromRequest (a RequestLog method) to read the
set the structs fields. Also, acquire the current time.

- Implement writeLogToFile to write the logData (of type RequestLog)
to a file in the tmp directory. Generate a unique uuid for the file
name, Marshal the struct into a json and then write it to the file.

- Implement loggingToFileMiddleware. Create the log from the Request
then call writeLogToFile(). In case of failure return
InternalServerError status code, else serve the next handler in the
chain.